+++ show/baby_del.post.php
@@ -1,14 +1,20 @@
 ﻿<?php
   require_once('../frame.php');
-	$id=$_REQUEST['id'];
-	$sql="delete from smg_baby_item where id=".$id;
-	$db = get_db();
-	if($db->execute($sql))
-	{
-		echo 'OK';
+  if(date('Y-m-d')!="2010-09-18"){
+		$id=$_REQUEST['id'];
+		$sql="delete from smg_baby_item where id=".$id;
+		$db = get_db();
+		if($db->execute($sql))
+		{
+			echo 'OK';
+		}
+		else
+		{
+			echo 'error';
+		}
 	}
 	else
 	{
-		echo 'error';
+		alert('对不起今天关闭一切提交！');	
 	}
 ?>
